Rename nav components to be more semantic and consistent
